make the form the index
